Change DatehourFromUnixtime to BIGINT

+++ plugin/trino-zettablock/src/main/java/io/trino/plugin/zettablock/udf/DatestrFunctions.java
@@ -52,7 +52,7 @@ public class DatestrFunctions
 
     @ScalarFunction("DatehourFromUnixtime")
     @Description("DatehourFromUnixtime")
-    @SqlType(StandardTypes.INTEGER)
+    @SqlType(StandardTypes.BIGINT)
     @SqlNullable
     public static Long datehourFromUnixtime(@SqlNullable @SqlType(StandardTypes.DOUBLE) Double unixtime)
     {
